requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
mne
scipy
pyedflib
numpy
